fix: base on main and rename package
